fix bad module names in extracted canvas files
canvas_probe_overlay.py and canvas_context_menu.py referenced
non-existent modules GUI.component_graphics and GUI.wire_graphics.
The correct names are GUI.component_item and GUI.wire_item.

These caused ModuleNotFoundError in CI, breaking all test files
that import circuit_canvas.py.
